% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_create_fleet}
\alias{robomaker_create_fleet}
\title{Creates a fleet, a logical group of robots running the same robot
application}
\usage{
robomaker_create_fleet(name, tags)
}
\arguments{
\item{name}{[required] The name of the fleet.}

\item{tags}{A map that contains tag keys and tag values that are attached to the
fleet.}
}
\description{
Creates a fleet, a logical group of robots running the same robot
application.
}
\section{Request syntax}{
\preformatted{robomaker$create_fleet(
  name = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
