% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robomaker_operations.R
\name{robomaker_update_robot_application}
\alias{robomaker_update_robot_application}
\title{Updates a robot application}
\usage{
robomaker_update_robot_application(application, sources,
  robotSoftwareSuite, currentRevisionId)
}
\arguments{
\item{application}{[required] The application information for the robot application.}

\item{sources}{[required] The sources of the robot application.}

\item{robotSoftwareSuite}{[required] The robot software suite used by the robot application.}

\item{currentRevisionId}{The revision id for the robot application.}
}
\description{
Updates a robot application.
}
\section{Request syntax}{
\preformatted{robomaker$update_robot_application(
  application = "string",
  sources = list(
    list(
      s3Bucket = "string",
      s3Key = "string",
      architecture = "X86_64"|"ARM64"|"ARMHF"
    )
  ),
  robotSoftwareSuite = list(
    name = "ROS",
    version = "Kinetic"
  ),
  currentRevisionId = "string"
)
}
}

\keyword{internal}
